LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-xmlhttprequest-redirect-to-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to 'http://localhost:8000/security/contentSecurityPolicy/resources/xhr-redirect-not-allowed.pl' because it violates the following Content Security Policy directive: "connect-src http://127.0.0.1:8000/security/contentSecurityPolicy/resources/redir.php".

PASS XMLHttpRequest.send() did not follow the disallowed redirect.
PASS successfullyParsed is true

TEST COMPLETE
